requirements.txt
# Core ML/AI packages
torch>=2.0.0
transformers>=4.57.0
accelerate>=1.12.0
bitsandbytes>=0.49.0
safetensors>=0.7.0
vllm>=0.6.0

# Web frameworks
gradio>=6.3.0
fastapi>=0.128.0
uvicorn>=0.40.0

# Image processing
pillow>=12.0.0

# HTTP client
requests>=2.32.0

# Standard dependencies
pydantic>=2.12.0
numpy>=2.4.0
